mysite/requirements.txt
certifi==2020.11.8
chardet==3.0.4
Django==2.2.17
idna==2.10
pinpointPy==1.0.0
PyMySQL
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
